+++ realm/realm-library/src/androidTest/java/io/realm/RealmTests.java
@@ -1160,8 +1160,7 @@ public class RealmTests {
         assertEquals(1, compactOnLaunchCount.get());
 
         realm = Realm.getInstance(realmConfig);
-        // Called 2 more times. The PK table migration logic (the old PK bug) needs to open/close the Realm once.
-        assertEquals(3, compactOnLaunchCount.get());
+        assertEquals(2, compactOnLaunchCount.get());
 
         Thread thread = new Thread(new Runnable() {
             @Override
@@ -1169,7 +1168,7 @@ public class RealmTests {
                 Realm bgRealm = Realm.getInstance(realmConfig);
                 bgRealm.close();
                 // compactOnLaunch should not be called anymore!
-                assertEquals(3, compactOnLaunchCount.get());
+                assertEquals(2, compactOnLaunchCount.get());
             }
         });
         thread.start();
@@ -1182,7 +1181,7 @@ public class RealmTests {
 
         realm.close();
 
-        assertEquals(3, compactOnLaunchCount.get());
+        assertEquals(2, compactOnLaunchCount.get());
     }
 
     @Test